Module teclado_helena::actions

Processamento de ações enviadas pelo cliente

Módulo para processar e simular ações do usuário como movimento do mouse, cliques, roda do mouse e teclas do teclado.

Enums

ButtonState

Representa o estado do botão de mouse ou da tecla do teclado

InputAction

Um enum representando todas as ações possíveis que o usuário pode enviar.

MouseButton

Representa os botões do mouse esquerdo e direito

ScrollDirection

Representa as possíveis direções de rolagem da roda do mouse

Functions

deserialize_action

Desserializa uma string JSON em um InputAction.

get_key 🔒

Converte uma representação em string de uma tecla em um [enigo::Key].

process_action

Processa uma ação de entrada do usuário

process_key_press

Simula um pressionamento de tecla

process_mouse_click

Simula uma ação de clique

process_mouse_move

Simula um movimento de mouse relativo

process_scroll

Processa uma ação de rolagem da roda do mouse